feat: rework config
im happy about this. this allows the config to be more expressive and the
implementation to be simpler. im happy with this, still could use a little sand
paper around the edges but the config feels pretty nice now. should definitely
figure out filtering by lights but we'll get there

Signed-off-by: Matt Provost <mattprovost6@gmail.com>